CORE-2634 Connection/reconnection fix.
This catches another serial failure mode, this time where select()
reports the device as ready to read, but no actual data is returned.
This previously manifested as simply being stuck in an unproductive
read loop, but now it correctly recovers from this state.

The slightly different ROS_ERROR message allowed me to verify this,
by re-plugging the device until I triggered this mode.